+++ .eslintrc.yml
@@ -1,3 +1,7 @@
+parserOptions:
+  sourceType: 'module'
+  ecmaVersion: 2015
+
 env:
   es6: true
   browser: true
