+++ .gitignore
@@ -1,15 +1,10 @@
-* text eol=lf
-
-*.png binary
-*.jpg binary
-*.jpeg binary
-*.webp binary
-*.woff binary
-*.woff2 binary
-
-.idea/
+*.log*
 node_modules/
-npm-debug.log*
-Thumbs.db
-.vscode/
+.idea
+
+# Папка с собранными файлами проекта
+build/
+# Собранные JS-файлы (бандл)
+public/bundle.js
+public/bundle.js.map
