+++ src/view/load-more-button.js
@@ -1,0 +1,5 @@
+export const createLoadMoreButtonTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
